add new file

+++ libs/Pwtree/Grant.php
@@ -11,10 +11,12 @@ class Pwtree_Grant{
 		  if(intval($v)<=0) return;
 	  }
 	  
-	  foreach(explode(",",$useridList) as $k=>$v){
-		  if ($v=='') continue;
-			if(intval($v)<=0) return;
-	  }
+	  if($groupid==''){
+		  foreach(explode(",",$useridList) as $k=>$v){
+			  if ($v=='') continue;
+				if(intval($v)<=0) return;
+		  }
+		}
 	  
 	  $conn =  Db_Mysqli::getIntance()->getConnection();
 	  
@@ -35,7 +37,6 @@ class Pwtree_Grant{
 	  $listArr = array();
 	 
 	  while ($stmt->fetch()) {
-		        
 				$listArr[$fid] = array("id"=>$fid,"treenavid"=>$treenavid,
 				"siteid"=>$siteid,"merid"=>$merid);
 	   }
@@ -44,10 +45,10 @@ class Pwtree_Grant{
 	  $stmt2='';
 	  $res2 ='';
 	  //var_dump($listArr);exit;
-	  if($listArr ){
+	 
 		$conn->autocommit(false);  
 		if($groupid!='' && $groupid>0){
-		    $delold = "delete from pw_permission_treenav where  f_groupid=? and f_merid = ? and f_siteid = ?";
+		  $delold = "delete from pw_permission_treenav where  f_groupid=? and f_merid = ? and f_siteid = ?";
 			$stmt2 = $conn->prepare($delold); 
 			$stmt2->bind_param('iii',$groupid,$merid,$siteid);		
 			$res2 = $stmt2->execute() ;	
@@ -57,44 +58,67 @@ class Pwtree_Grant{
 				  SeasLog::log(SEASLOG_ERROR,mysqli_error($conn));
 				  return false;
 				}
-		}		
-         	
-		foreach(explode(",",$useridList) as $uk=>$uv){
-			 if(intval($uv)<=0)continue;			  
-			 $delold = '';		 
-			 $delold = "delete from pw_permission_treenav where  f_userid=? and f_merid = ? and f_siteid = ?";
-			 $stmt2 = $conn->prepare($delold); 
-			 $stmt2->bind_param('iii',$uv,$merid,$siteid);		
-			 $res2 = $stmt2->execute() ;	
-			 if($res2==false)
+			
+			$treesql = "insert into pw_permission_treenav(f_userid,f_treenavid,f_siteid,f_permissionid,f_merid,f_groupid)values(?,?,?,?,?,?)";
+			foreach($temp_pemidList as $k=>$v){	 
+			$stmt2 = $conn->prepare($treesql); 
+			if(array_key_exists($v,$listArr)){
+				 $stmt2->bind_param('iiiiii',$userid=0 ,$listArr[$v]['treenavid'],$siteid,$listArr[$v]['id'],$merid,$groupid);
+				}else{			 
+					$stmt2->bind_param('iiiiii',$userid=0 ,$v,$siteid,$pid=-1,$merid,$groupid);	
+				}
+				$res2 = $stmt2->execute() ;	
+		       //var_dump($res2);
+			  if($res2==false)
 				{
 				  $conn->rollback();
 				  SeasLog::log(SEASLOG_ERROR,mysqli_error($conn));
-				  return false;
-				}
-			
-		 foreach($temp_pemidList as $k=>$v){	   
+				  return;
+				 }
+				} 		
+				
+		}else if ($groupid=='' &&  $useridList !=''){	
+			foreach(explode(",",$useridList) as $uk=>$uv){
+				 if(intval($uv)<=0)continue;			  
+				 $delold = '';		 
+				 $delold = "delete from pw_permission_treenav where  f_userid=? and f_merid = ? and f_siteid = ?";
+				 $stmt2 = $conn->prepare($delold); 
+				 $stmt2->bind_param('iii',$uv,$merid,$siteid);		
+				 $res2 = $stmt2->execute() ;	
+				 if($res2==false)
+					{
+					  $conn->rollback();
+					  SeasLog::log(SEASLOG_ERROR,mysqli_error($conn));
+					  return false;
+					}
+				
 			$treesql = "insert into pw_permission_treenav(f_userid,f_treenavid,f_siteid,f_permissionid,f_merid,f_groupid)values(?,?,?,?,?,?)";
-			$stmt2 = $conn->prepare($treesql); 
-			
-	        if(array_key_exists($v,$listArr)){
-			 $stmt2->bind_param('iiiiii',$uv ,$listArr[$v]['treenavid'],$siteid,$listArr[$v]['id'],$merid,$groupid);
-			}else{			 
-				$stmt2->bind_param('iiiiii',$uv ,$v,$siteid,$pid=-1,$merid,$groupid);	
-			}
-			$res2 = $stmt2->execute() ;	
-	       //var_dump($res2);
-		   if($res2==false)
-			{
-			  $conn->rollback();
-			  SeasLog::log(SEASLOG_ERROR,mysqli_error($conn));
-			  return;
-			} 
-		   }
-		 }		
-	   }
+			 foreach($temp_pemidList as $k=>$v){	   				
+				$stmt2 = $conn->prepare($treesql); 
+				
+		    if(array_key_exists($v,$listArr)){
+				 $stmt2->bind_param('iiiiii',$uv ,$listArr[$v]['treenavid'],$siteid,$listArr[$v]['id'],$merid,$groupid);
+				}else{			 
+					$stmt2->bind_param('iiiiii',$uv ,$v,$siteid,$pid=-1,$merid,$groupid);	
+				}
+				$res2 = $stmt2->execute() ;	
+		       //var_dump($res2);
+			   if($res2==false)
+				{
+				  $conn->rollback();
+				  SeasLog::log(SEASLOG_ERROR,mysqli_error($conn));
+				  return;
+				 } 
+			   }
+			 }		
+	    }
+	    
+	    
+	  
+	  
+	  
 		$conn->commit();
-        if($stmt2){		$stmt2->close();}
+    if($stmt2){		$stmt2->close();}
 		return $res2;
 	}
